Move NF3 actions class into /extras subdir
